v 1.0.0
A repository containing the following list of items:
1. Different types of algorithms within some common algorithmic domains necessary to master.
2. Common and not so common algorithms for different data structures.
3. Solutions to questions from some of the most top competitive coding websites.
4. Company specific problems.
5. Some extra codes that are part of some online MOOCs or a textbooks based on Algorithms and Data Structure.

LICENSE
Copyright 2017 Aayush Goyal

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.